+++ src/lib/Data/Geolocation/GEOS/Trans.hs
@@ -1,14 +1,14 @@
 {-|
 Module      : Data.Geolocation.GEOS.Trans
-Description : @MaybeT@ wrappers for high-level GEOS API
+Description : Monad transformer wrappers for high-level GEOS API
 Copyright   : (C) Richard Cook, 2016
 Licence     : MIT
 Maintainer  : rcook@rcook.org
 Stability   : experimental
 Portability : portable
 
-These are @MaybeT@ monad transformer wrapper functions for the high-level API
-to simplify error handling in client code.
+These are monad transformer wrapper functions for the high-level API to
+simplify error handling in client code.
 
 For the low-level FFI bindings, see "Data.Geolocation.GEOS.Imports".
 
@@ -18,7 +18,9 @@ For the high-level API, see "Data.Geolocation.GEOS".
 -}
 
 module Data.Geolocation.GEOS.Trans
-    ( areaM
+    ( GEOSError (..)
+    , MonadGEOS
+    , areaM
     , createCollectionM
     , createCoordSeqM
     , createEmptyPolygonM
@@ -41,7 +43,6 @@ module Data.Geolocation.GEOS.Trans
     , mkWriterM
     , readGeometryM
     , runGEOS
-    , runGEOSEither
     , setOrdinateM
     , setXM
     , setYM
@@ -50,158 +51,152 @@ module Data.Geolocation.GEOS.Trans
     ) where
 
 import Control.Applicative
-import Control.Monad.Trans.Maybe
+import Control.Monad.Trans.Class
+import Control.Monad.Trans.Except
 import Data.Geolocation.GEOS
 import Data.Word
 
-unaryGEOSFunc :: (a -> IO (Maybe b)) -> a -> MaybeT IO b
-unaryGEOSFunc = (MaybeT .)
+data GEOSError = GEOSError String deriving Show
 
-binaryGEOSFunc :: (a -> b -> IO (Maybe c)) -> a -> b -> MaybeT IO c
-binaryGEOSFunc f a b = MaybeT (f a b)
+note :: IO (Maybe a) -> IO (Either GEOSError a)
+note action = do
+    result <- action
+    case result of
+         Nothing -> getErrorMessage >>= return . Left . GEOSError
+         Just value -> return $ Right value
 
-ternaryGEOSFunc :: (a -> b -> c -> IO (Maybe d)) -> a -> b -> c -> MaybeT IO d
-ternaryGEOSFunc f a b c = MaybeT (f a b c)
+type MonadGEOS a = ExceptT GEOSError IO a
 
-quaternaryGEOSFunc :: (a -> b -> c -> d -> IO (Maybe e)) -> a -> b -> c -> d -> MaybeT IO e
-quaternaryGEOSFunc f a b c d = MaybeT (f a b c d)
+unaryGEOSFunc :: (a -> IO (Maybe b)) -> a -> MonadGEOS b
+unaryGEOSFunc f a = ExceptT $ note (f a)
 
--- |@MaybeT@-wrapped version of 'area'
-areaM :: Geometry -> MaybeT IO Double
+binaryGEOSFunc :: (a -> b -> IO (Maybe c)) -> a -> b -> MonadGEOS c
+binaryGEOSFunc f a b = ExceptT $ note (f a b)
+
+ternaryGEOSFunc :: (a -> b -> c -> IO (Maybe d)) -> a -> b -> c -> MonadGEOS d
+ternaryGEOSFunc f a b c = ExceptT $ note (f a b c)
+
+quaternaryGEOSFunc :: (a -> b -> c -> d -> IO (Maybe e)) -> a -> b -> c -> d -> MonadGEOS e
+quaternaryGEOSFunc f a b c d = ExceptT $ note (f a b c d)
+
+-- |@MonadGEOS@-wrapped version of 'area'
+areaM :: Geometry -> MonadGEOS Double
 areaM = unaryGEOSFunc area
 
--- |@MaybeT@-wrapped version of 'createCollection'
-createCollectionM :: GeometryType -> [Geometry] -> MaybeT IO Geometry
+-- |@MonadGEOS@-wrapped version of 'createCollection'
+createCollectionM :: GeometryType -> [Geometry] -> MonadGEOS Geometry
 createCollectionM = binaryGEOSFunc createCollection
 
--- |@MaybeT@-wrapped version of 'createCoordSeq'
-createCoordSeqM :: Context -> Word -> Word -> MaybeT IO CoordinateSequence
+-- |@MonadGEOS@-wrapped version of 'createCoordSeq'
+createCoordSeqM :: Context -> Word -> Word -> MonadGEOS CoordinateSequence
 createCoordSeqM = ternaryGEOSFunc createCoordSeq
 
--- |@MaybeT@-wrapped version of 'createEmptyPolygon'
-createEmptyPolygonM :: Context -> MaybeT IO Geometry
+-- |@MonadGEOS@-wrapped version of 'createEmptyPolygon'
+createEmptyPolygonM :: Context -> MonadGEOS Geometry
 createEmptyPolygonM = unaryGEOSFunc createEmptyPolygon
 
--- |@MaybeT@-wrapped version of 'createLinearRing'
-createLinearRingM :: CoordinateSequence -> MaybeT IO Geometry
+-- |@MonadGEOS@-wrapped version of 'createLinearRing'
+createLinearRingM :: CoordinateSequence -> MonadGEOS Geometry
 createLinearRingM = unaryGEOSFunc createLinearRing
 
--- |@MaybeT@-wrapped version of 'createPolygon'
-createPolygonM :: Geometry -> [Geometry] -> MaybeT IO Geometry
+-- |@MonadGEOS@-wrapped version of 'createPolygon'
+createPolygonM :: Geometry -> [Geometry] -> MonadGEOS Geometry
 createPolygonM = binaryGEOSFunc createPolygon
 
--- |@MaybeT@-wrapped version of 'envelope'
-envelopeM :: Geometry -> MaybeT IO Geometry
+-- |@MonadGEOS@-wrapped version of 'envelope'
+envelopeM :: Geometry -> MonadGEOS Geometry
 envelopeM = unaryGEOSFunc envelope
 
--- |@MaybeT@-wrapped version of 'geomTypeId'
-geomTypeIdM :: Geometry -> MaybeT IO GeometryType
+-- |@MonadGEOS@-wrapped version of 'geomTypeId'
+geomTypeIdM :: Geometry -> MonadGEOS GeometryType
 geomTypeIdM = unaryGEOSFunc geomTypeId
 
--- |@MaybeT@-wrapped version of 'getCoordSeq'
-getCoordSeqM :: Geometry -> MaybeT IO CoordinateSequence
+-- |@MonadGEOS@-wrapped version of 'getCoordSeq'
+getCoordSeqM :: Geometry -> MonadGEOS CoordinateSequence
 getCoordSeqM = unaryGEOSFunc getCoordSeq
 
--- |@MaybeT@-wrapped version of 'getExteriorRing'
-getExteriorRingM :: Geometry -> MaybeT IO Geometry
+-- |@MonadGEOS@-wrapped version of 'getExteriorRing'
+getExteriorRingM :: Geometry -> MonadGEOS Geometry
 getExteriorRingM = unaryGEOSFunc getExteriorRing
 
--- |@MaybeT@-wrapped version of 'getGeometry'
-getGeometryM :: Geometry -> Int -> MaybeT IO Geometry
+-- |@MonadGEOS@-wrapped version of 'getGeometry'
+getGeometryM :: Geometry -> Int -> MonadGEOS Geometry
 getGeometryM = binaryGEOSFunc getGeometry
 
--- |@MaybeT@-wrapped version of 'getNumGeometries'
-getNumGeometriesM :: Geometry -> MaybeT IO Int
+-- |@MonadGEOS@-wrapped version of 'getNumGeometries'
+getNumGeometriesM :: Geometry -> MonadGEOS Int
 getNumGeometriesM = unaryGEOSFunc getNumGeometries
 
--- |@MaybeT@-wrapped version of 'getOrdinate'
-getOrdinateM :: CoordinateSequence -> Word -> Word -> MaybeT IO Double
+-- |@MonadGEOS@-wrapped version of 'getOrdinate'
+getOrdinateM :: CoordinateSequence -> Word -> Word -> MonadGEOS Double
 getOrdinateM = ternaryGEOSFunc getOrdinate
 
--- |@MaybeT@-wrapped version of 'getSize'
-getSizeM :: CoordinateSequence -> MaybeT IO Word
+-- |@MonadGEOS@-wrapped version of 'getSize'
+getSizeM :: CoordinateSequence -> MonadGEOS Word
 getSizeM = unaryGEOSFunc getSize
 
--- |@MaybeT@-wrapped version of 'getX'
-getXM :: CoordinateSequence -> Word -> MaybeT IO Double
+-- |@MonadGEOS@-wrapped version of 'getX'
+getXM :: CoordinateSequence -> Word -> MonadGEOS Double
 getXM = binaryGEOSFunc getX
 
--- |@MaybeT@-wrapped version of 'getY'
-getYM :: CoordinateSequence -> Word -> MaybeT IO Double
+-- |@MonadGEOS@-wrapped version of 'getY'
+getYM :: CoordinateSequence -> Word -> MonadGEOS Double
 getYM = binaryGEOSFunc getY
 
--- |@MaybeT@-wrapped version of 'getZ'
-getZM :: CoordinateSequence -> Word -> MaybeT IO Double
+-- |@MonadGEOS@-wrapped version of 'getZ'
+getZM :: CoordinateSequence -> Word -> MonadGEOS Double
 getZM = binaryGEOSFunc getZ
 
--- |@MaybeT@-wrapped version of 'intersection'
-intersectionM :: Geometry -> Geometry -> MaybeT IO Geometry
+-- |@MonadGEOS@-wrapped version of 'intersection'
+intersectionM :: Geometry -> Geometry -> MonadGEOS Geometry
 intersectionM = binaryGEOSFunc intersection
 
--- |@MaybeT@-wrapped version of 'isEmpty'
-isEmptyM :: Geometry -> MaybeT IO Bool
+-- |@MonadGEOS@-wrapped version of 'isEmpty'
+isEmptyM :: Geometry -> MonadGEOS Bool
 isEmptyM = unaryGEOSFunc isEmpty
 
--- |@MaybeT@-wrapped version of 'mkReader'
-mkReaderM :: Context -> MaybeT IO Reader
+-- |@MonadGEOS@-wrapped version of 'mkReader'
+mkReaderM :: Context -> MonadGEOS Reader
 mkReaderM = unaryGEOSFunc mkReader
 
--- |@MaybeT@-wrapped version of 'mkWriter'
-mkWriterM :: Context -> MaybeT IO Writer
+-- |@MonadGEOS@-wrapped version of 'mkWriter'
+mkWriterM :: Context -> MonadGEOS Writer
 mkWriterM = unaryGEOSFunc mkWriter
 
--- |@MaybeT@-wrapped version of 'readGeometry'
-readGeometryM :: Reader -> String -> MaybeT IO Geometry
+-- |@MonadGEOS@-wrapped version of 'readGeometry'
+readGeometryM :: Reader -> String -> MonadGEOS Geometry
 readGeometryM = binaryGEOSFunc readGeometry
 
 -- |Creates a <https://trac.osgeo.org/geos/ GEOS> context, passes it to a block
 -- and releases the context and all associated objects such as readers, writers
--- and geometries at the end inside a @MaybeT IO@ monad:
---
--- @
---    runGEOS $ \ctx -> do
---
---        -- Use context
---
---        return ()
--- @
-runGEOS :: (Context -> MaybeT IO a) -> IO (Maybe a)
-runGEOS = withGEOS . (runMaybeT .)
-
--- |Creates a <https://trac.osgeo.org/geos/ GEOS> context, passes it to a block
--- and releases the context and all associated objects such as readers, writers
--- and geometries at the end inside a @MaybeT IO@ monad returning a message in
+-- and geometries at the end inside a @MonadGEOS@ monad returning a message in
 -- case of error:
 --
 -- @
---    result <- runGEOSEither $ \ctx -> do
+--    result <- runGEOS $ \ctx -> do
 --
 --        -- Use context
 --
 --        return ()
 --    case result of
---          Left m -> putStrLn $ "Failed: " ++ m
+--          Left (GEOSError m) -> putStrLn $ "Failed: " ++ m
 --          Right r -> putStrLn $ "Succeeded: " ++ show r
 -- @
-runGEOSEither :: (Context -> MaybeT IO a) -> IO (Either String a)
-runGEOSEither action = do
-    result <- runGEOS action
-    case result of
-         Nothing -> Left <$> getErrorMessage
-         Just r -> return $ Right r
+runGEOS :: (Context -> MonadGEOS a) -> IO (Either GEOSError a)
+runGEOS = withGEOS . (runExceptT .)
 
--- |@MaybeT@-wrapped version of 'setOrdinate'
+-- |@MonadGEOS@-wrapped version of 'setOrdinate'
 setOrdinateM = quaternaryGEOSFunc setOrdinate
 
--- |@MaybeT@-wrapped version of 'setX'
+-- |@MonadGEOS@-wrapped version of 'setX'
 setXM = ternaryGEOSFunc setX
 
--- |@MaybeT@-wrapped version of 'setY'
+-- |@MonadGEOS@-wrapped version of 'setY'
 setYM = ternaryGEOSFunc setY
 
--- |@MaybeT@-wrapped version of 'setZ'
+-- |@MonadGEOS@-wrapped version of 'setZ'
 setZM = ternaryGEOSFunc setZ
 
--- |@MaybeT@-wrapped version of 'area'
-writeGeometryM :: Writer -> Geometry -> MaybeT IO String
+-- |@MonadGEOS@-wrapped version of 'area'
+writeGeometryM :: Writer -> Geometry -> MonadGEOS String
 writeGeometryM = binaryGEOSFunc writeGeometry